Moved account events into separate directory
no-issue

This is just a tidying because the account directory is getting full.
