+++ src/account/events/index.ts
@@ -1,0 +1,8 @@
+export { AccountBlockedEvent } from './account-blocked.event';
+export { AccountFollowedEvent } from './account-followed.event';
+export { AccountUnblockedEvent } from './account-unblocked.event';
+export { AccountUnfollowedEvent } from './account-unfollowed.event';
+export { AccountUpdatedEvent } from './account-updated.event';
+export { DomainBlockedEvent } from './domain-blocked.event';
+export { DomainUnblockedEvent } from './domain-unblocked.event';
+export { NotificationsReadEvent } from './notifications-read-event';
