+++ src/feed/feed-update.service.unit.test.ts
@@ -2,10 +2,12 @@ import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { EventEmitter } from 'node:events';
 
-import { AccountBlockedEvent } from 'account/account-blocked.event';
-import { AccountUnfollowedEvent } from 'account/account-unfollowed.event';
 import { AccountEntity } from 'account/account.entity';
-import { DomainBlockedEvent } from 'account/domain-blocked.event';
+import {
+    AccountBlockedEvent,
+    AccountUnfollowedEvent,
+    DomainBlockedEvent,
+} from 'account/events';
 import { FeedUpdateService } from 'feed/feed-update.service';
 import type { FeedService } from 'feed/feed.service';
 import { PostCreatedEvent } from 'post/post-created.event';
